apps/cloud-functions/resume_to_json_v1/requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
Flask==3.0.0
functions-framework==3.5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pdf2image==1.16.3
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
